test: remove circular dep
