+++ packages/browser/src/profiling/hubextensions.ts
@@ -4,14 +4,14 @@ import type { Transaction } from '@sentry/types';
 import { logger, uuid4 } from '@sentry/utils';
 
 import { WINDOW } from '../helpers';
-import { addToProfileQueue } from './integration';
 import type {
   JSSelfProfile,
   JSSelfProfiler,
   JSSelfProfilerConstructor,
   ProcessedJSSelfProfile,
 } from './jsSelfProfiling';
-import { isValidSampleRate } from './utils';
+import { addToProfileQueue , isValidSampleRate } from './utils';
+
 
 export const MAX_PROFILE_DURATION_MS = 30_000;
 // Keep a flag value to avoid re-initializing the profiler constructor. If it fails
@@ -270,3 +270,5 @@ export function wrapTransactionWithProfiling(transaction: Transaction): Transact
   transaction.finish = profilingWrappedTransactionFinish;
   return transaction;
 }
+
+
